% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockCoefficients.R
\name{getClockCoefficients}
\alias{getClockCoefficients}
\title{Access Packaged Clock Coefficients}
\usage{
getClockCoefficients()
}
\value{
A named list of data frames. Each data frame contains the
coefficients for one clock model.
}
\description{
Returns the packaged list of pre-trained cell-type specific
clock coefficient tables included with CellAgeR. Available clock names
are conceptually structured by tissue and cell type:
Brain (\code{NeuronClock}, \code{sortedNeuronClock}, \code{OligoClock}, \code{sortedOligoClock});
Epithelial (\code{BreastEpiClock}, \code{ColonEpiClock}, \code{KeraClock}, \code{LungEpiClock},
\code{ProstateEpiClock}, \code{salivaEpiClock}); Fibroblast (\code{BreastFibClock},
\code{ColonFibClock}, \code{LungFibClock}); Immune (\code{Cd4tClock}, \code{ColonImmunClock},
\code{LungImmunClock}, \code{MonoClock}, \code{salivaImmunClock}); and Liver (\code{HepClock}).
}
\examples{
names(getClockCoefficients())
}
