% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCTSAge.R
\name{predictCTSAge}
\alias{predictCTSAge}
\title{Predict Epigenetic Age and Age Acceleration}
\usage{
predictCTSAge(betaMatrix, clock, age = NULL)
}
\arguments{
\item{betaMatrix}{A numeric matrix of beta values, with
CpGs in rows and samples in columns.}

\item{clock}{A data frame representing the epigenetic clock model. It must
contain 'Feature' and 'Coefficient' columns. The intercept should be
indicated by '(Intercept)' in the 'Feature' column.}

\item{age}{A numeric vector of chronological ages for the samples. The
order must correspond to the sample order in \code{betaMatrix}. Defaults
to NULL.}
}
\value{
A list containing the following components:
\item{predictions}{A data frame with the predicted DNAmAge for each sample.}
\item{ageAcceleration}{A data frame with Age Acceleration Residuals for
each sample. NULL if \code{age} is not provided.}
}
\description{
This function predicts epigenetic age based on a given DNA
methylation beta matrix and a clock model. Optionally, if chronological
age is provided, it calculates age acceleration residuals.
}
