% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockCoefficients.R
\name{getClockCoefficients}
\alias{getClockCoefficients}
\title{Access Packaged Clock Coefficients}
\usage{
getClockCoefficients()
}
\value{
A named list of data frames. Each data frame contains the
coefficients for one clock model.
}
\description{
Returns the packaged list of pre-trained cell-type specific
clock coefficient tables included with CellAgeR. Available clock names
are \code{ProstateEpiClock}, \code{NeuronClock}, \code{ColonImmunClock}, \code{Cd4tClock},
\code{sortedOligoClock}, \code{ColonEpiClock}, \code{sortedNeuronClock}, \code{LungFibClock},
\code{KeraClock}, \code{BreastFibClock}, \code{salivaImmunClock}, \code{LungEpiClock},
\code{BreastEpiClock}, \code{salivaEpiClock}, \code{MonoClock}, \code{HepClock},
\code{ColonFibClock}, \code{OligoClock}, and \code{LungImmunClock}.
}
\examples{
names(getClockCoefficients())
}
