% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCTSAge.R
\name{predictCTSAge}
\alias{predictCTSAge}
\title{Predict Epigenetic Age and Age Acceleration}
\usage{
predictCTSAge(betaMatrix, clock, age = NULL)
}
\arguments{
\item{betaMatrix}{A numeric matrix of beta values, with
CpGs in rows and samples in columns.}

\item{clock}{A data frame representing the epigenetic clock model. It must
contain 'Feature' and 'Coefficient' columns. The intercept should be
indicated by '(Intercept)' in the 'Feature' column. If you use packaged
models from \code{getClockCoefficients()}, valid clock names are conceptually
structured by tissue and cell type: Brain (\code{NeuronClock}, \code{sortedNeuronClock},
\code{OligoClock}, \code{sortedOligoClock}); Epithelial (\code{BreastEpiClock}, \code{ColonEpiClock},
\code{KeraClock}, \code{LungEpiClock}, \code{ProstateEpiClock}, \code{salivaEpiClock}); Fibroblast
(\code{BreastFibClock}, \code{ColonFibClock}, \code{LungFibClock}); Immune (\code{Cd4tClock},
\code{ColonImmunClock}, \code{LungImmunClock}, \code{MonoClock}, \code{salivaImmunClock}); and Liver
(\code{HepClock}). For example, use
\code{clock = getClockCoefficients()$NeuronClock}.}

\item{age}{A numeric vector of chronological ages for the samples. The
order must correspond to the sample order in \code{betaMatrix}. Defaults
to NULL.}
}
\value{
A list containing the following components:
\item{predictions}{A data frame with the predicted DNAmAge for each sample.}
\item{ageAcceleration}{A data frame with Age Acceleration Residuals for
each sample. NULL if \code{age} is not provided.}
}
\description{
This function predicts epigenetic age based on a given DNA
methylation beta matrix and a clock model. Optionally, if chronological
age is provided, it calculates age acceleration residuals. When using
packaged models from \code{getClockCoefficients()}, available clock names are
conceptually structured by tissue and cell type: Brain (\code{NeuronClock},
\code{sortedNeuronClock}, \code{OligoClock}, \code{sortedOligoClock}); Epithelial
(\code{BreastEpiClock}, \code{ColonEpiClock}, \code{KeraClock}, \code{LungEpiClock},
\code{ProstateEpiClock}, \code{salivaEpiClock}); Fibroblast (\code{BreastFibClock},
\code{ColonFibClock}, \code{LungFibClock}); Immune (\code{Cd4tClock}, \code{ColonImmunClock},
\code{LungImmunClock}, \code{MonoClock}, \code{salivaImmunClock}); and Liver (\code{HepClock}).
}
\examples{
clock_coefficients <- getClockCoefficients()
example_dir <- system.file("extdata", package = "CellAgeR")
load(file.path(example_dir, "Walker.Rdata"))

res <- predictCTSAge(
  betaMatrix = as.matrix(betaMat[, 1:3]),
  clock = clock_coefficients$NeuronClock,
  age = phenoDf$Age[1:3]
)

head(res$predictions)
head(res$ageAcceleration)

}
